djangoAPI/requirements.txt
asgiref==3.5.2
Django==4.0.6
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
django-filter==22.1
gunicorn==20.1.0